backend/models/generator.py
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Conv2D, Flatten, Reshape, UpSampling2D
from tensorflow.keras.optimizers import Adam

class AxolotlGenerator:
    def __init__(self, input_shape=(100,), img_shape=(64, 64, 3)):
        self.input_shape = input_shape
        self.img_shape = img_shape
        self.model = self.build_model()

    def build_model(self):
        model = Sequential()
        model.add(Dense(256, input_dim=self.input_shape[0]))
        model.add(Reshape((4, 4, 16)))
        model.add(UpSampling2D())
        model.add(Conv2D(128, kernel_size=3, padding='same', activation='relu'))
        model.add(UpSampling2D())
        model.add(Conv2D(64, kernel_size=3, padding='same', activation='relu'))
        model.add(Conv2D(3, kernel_size=3, padding='same', activation='tanh'))
        model.compile(loss='binary_crossentropy', optimizer=Adam())
        return model

    def generate_image(self, noise):
        generated_image = self.model.predict(noise)
        return generated_image

    def train(self, dataset, epochs=10000, batch_size=32):
        for epoch in range(epochs):
            noise = np.random.normal(0, 1, (batch_size, self.input_shape[0]))
            generated_images = self.generate_image(noise)

            # Here you would add code to train the model on the dataset
            # This is a placeholder for the training process
            # model.train_on_batch(real_images, generated_images) 

            if epoch % 1000 == 0:
                print(f'Epoch {epoch} completed')