backend/requirements.txt
Flask
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
Pillow>=9.0.0
requests>=2.25.0
flask-cors>=3.0.10
icrawler>=0.6.0
pytesseract>=0.3.8
tqdm>=4.60.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
scikit-image>=0.18.0